Purge Report GlobalProperty
When saving a report definition, a global property is used to track the
version of the report. Using the reportUtil function to setup reports on
module init the version gets checked, preventing install if the versions
are the same even if there is no definition attatched to the version.
So no purge the global prop when purging the report def.

MEPTS-24