Add age dimensions for TX_NEW
The age disaggregation was implemented inside the SQL for the TX_NEW,
this code was removed in favour of using an age dimension in the dataset.

Sinse it is not possible to pass the patient ART start date as a parameter
for AgeCohortDefinitions custom CalculationCohortDefinitions were implemented
for determining if the age in within a certain interval on ART start date.