Fix Avoid unused private fields such as 'dao'
Implemented integration test for EptsReportsDao for this it was necessary to add
`TestingApplicationContext.xml` to be able to use hibernate mapping for table
`report_design`.

One of the mappings uses type `org.openmrs.module.reporting.report.service.db.PropertiesType`
which is only available in the reporting module `api-1.9` and `api-2.0` so `reporting-api-2.0`
was added as dependency.